Fixes for Travis
Fixes for Travis
