--- conftest.py
@@ -1,26 +1,0 @@
-import pytest
-from django.conf import settings
-from django.test import RequestFactory
-
-
-import os
-import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '')))
-
-
-from tests.factories import UserFactory
-
-
-@pytest.fixture(autouse=True)
-def media_storage(settings, tmpdir):
-    settings.MEDIA_ROOT = tmpdir.strpath
-
-
-@pytest.fixture
-def user() -> settings.AUTH_USER_MODEL:
-    return UserFactory()
-
-
-@pytest.fixture
-def request_factory() -> RequestFactory:
-    return RequestFactory()
