+++ pytest.ini
@@ -1,6 +1,7 @@
 [pytest]
 addopts =
+    --cov=csci_2019_project
     --cov-branch
     --pyargs
-
+testpaths = csci_2019_project
 DJANGO_SETTINGS_MODULE=config.settings.test
